Change order of target to draw metric last
Currently the graph line for the metric is drawn first, followed by that
for the various attributes such as warning, critical, min and max. In
cases where the metric follows the value of an attribute exactly, only the
attribute is seen because it is drawn later. This can make it appear that the
metric is not there!

This change places the metric last so that it will always be visible. A
side effect is that the metric is also listed last in the legend.

If this is undesirable for some, I suggest introducing a configuration
option to allow a choice.